+++ test/gurka/test_trace_attributes.cc
@@ -364,3 +364,32 @@ TEST(Standalone, RetrieveEdgeTrafficSignal) {
   EXPECT_TRUE(edges[1].HasMember("traffic_signal"));
   EXPECT_FALSE(edges[1]["traffic_signal"].GetBool());
 }
+
+TEST(Standalone, ViaFerrataNoSacScale) {
+  const std::string ascii_map = R"(
+    A--B--C
+  )";
+
+  const gurka::ways ways = {{"AB", {{"highway", "via_ferrata"}}},
+                            {"BC", {{"highway", "via_ferrata"}}}};
+
+  const double gridsize = 10;
+  const auto layout = gurka::detail::map_to_coordinates(ascii_map, gridsize);
+  auto map = gurka::buildtiles(layout, ways, {}, {}, "test/data/sac_scale_attributes");
+
+  std::string trace_json;
+  auto api = gurka::do_action(valhalla::Options::trace_attributes, map, {"A", "B", "C"}, "pedestrian",
+                              {{"/costing_options/pedestrian/max_hiking_difficulty", "6"}}, {},
+                              &trace_json, "via");
+
+  rapidjson::Document result;
+  result.Parse(trace_json.c_str());
+
+  auto edges = result["edges"].GetArray();
+  ASSERT_EQ(edges.Size(), 2);
+
+  EXPECT_TRUE(edges[0].HasMember("sac_scale"));
+  EXPECT_EQ(edges[0]["sac_scale"].GetInt(), 6);
+  EXPECT_TRUE(edges[1].HasMember("sac_scale"));
+  EXPECT_EQ(edges[1]["sac_scale"].GetInt(), 6);
+}